feat: [service-runtime-state-consolidation-recovery-007] - Land S2 on main reconciled with post-S1 main
Resolve pkg-maint violations so the recovery branch is CI merge-ready after reconciling with post-S1 main.
