+++ pkg/service/factory_build.go
@@ -235,6 +235,22 @@ func buildRuntimeBundle(
 	eventHistory := factoryevents.NewFactoryEventHistory(net, input.clock.Now, input.loadedFactoryCfg)
 	eventHistory.SetFactoryRunnerOverride(effectiveFactoryRunnerID)
 	localModels := newRuntimeLocalModelDependencies(input.cfg)
+	workerOpts, err := loadRuntimeBundleWorkerOptions(input, logger, effectiveFactoryRunnerID, eventHistory, localModels)
+	if err != nil {
+		return nil, err
+	}
+
+	bundleBuilt = true
+	return assembleRuntimeBundle(input, logger, logSink, net, eventHistory, localModels, workerOpts)
+}
+
+func loadRuntimeBundleWorkerOptions(
+	input runtimeBundleBuildInput,
+	logger *zap.Logger,
+	effectiveFactoryRunnerID string,
+	eventHistory *factoryevents.FactoryEventHistory,
+	localModels localModelDomain,
+) ([]factory.FactoryOption, error) {
 	workerOpts, err := loadWorkersFromConfig(
 		input.loadedFactoryCfg.FactoryDir(),
 		input.loadedFactoryCfg.FactoryConfig(),
@@ -256,7 +272,18 @@ func buildRuntimeBundle(
 		logger.Error("failed to load workers from config", zap.Error(err))
 		return nil, fmt.Errorf("load workers: %w", err)
 	}
+	return workerOpts, nil
+}
 
+func assembleRuntimeBundle(
+	input runtimeBundleBuildInput,
+	logger *zap.Logger,
+	logSink *logging.RuntimeLogSink,
+	net *state.Net,
+	eventHistory *factoryevents.FactoryEventHistory,
+	localModels localModelDomain,
+	workerOpts []factory.FactoryOption,
+) (*factoryRuntimeBundle, error) {
 	recording, err := buildRuntimeRecorder(
 		input.cfg,
 		input.loadedFactoryCfg.FactoryDir(),
@@ -299,7 +326,6 @@ func buildRuntimeBundle(
 		return nil, err
 	}
 
-	bundleBuilt = true
 	return &factoryRuntimeBundle{
 		dir:            input.dir,
 		folderPath:     input.folderPath,
